Implement k_thread_user_mode_enter in safe Rust
A good exercise in passing a closure to a C function that wants a
callback. The unreachable!() and call to panic that it generates would
not be necessary if the bindings flagged that the zephyr user_mode_enter
does not return.

This is also unsafe in the sense that it can be called when already in
user mode. Having a token to track the context at compile time would be
better. For now it can crash if used incorrectly.