Generate thunks and bindings for inline syscalls
Both with and without CONFIG_USERSPACE, generate thunks to the z_impl_*
kernel syscalls as z_kernelctx_*, similar to the "user" and "any"
context thunks. These just directly pass through to the kernel function.

This gives us a concrete implementation of all inline functions and an
unnecessary tail-call to functions that already have a concrete
implementation. In zephyr-bindgen we can get the set of non-inline
z_impl functions because bindgen doesn't generate bindings for inlines.
Having the set of all syscalls and the set of non-inline kernel impls
allows us to generate the set of inlines and point rust at our thunks
for only those. The others will be garbage collected.