Put Rust data/bss in app shared memory domain
Link an object from all of the Rust sources which allows using a linker
script to put data/bss from Rust in an app memory region.

All the globals used in libstd and any Rust libraries need to be
accessible by all Rust threads, so we put them in one memory domain
along with a mem pool to serve as a global allocator.

Currently still using the mem pool if !usermode. Could use kmalloc
instead. Also the pool is created in rust-smem.c but must be set as the
global allocator in the top level app crate. Need to somehow tie the
creation and use of those together.