Add test
Symlink the rust-app so that we can build it differently using a
different Cargo.toml. The only difference is that we include the feature
"have_ztest" in rust-app/zephyr to pull in the ztest_test_fail()
method. Technically this is not required; it simply allows a failing
test to exist cleanly rather than panicing.