UART FIFO-buffered driver and Rust bindings
Allows wrapping the raw UART API with software RX and TX FIFOs that can
be accessed from user space. Interrupt handlers move data between the
hardware fifos and memory. This way, interrupt handlers do not need to
be specifically written to deal with application-specific UART use cases
from kernel mode. Signals are provided for a thread to be woken on read
or write readiness. Behaves similarly to a UNIX file, supporting
blocking and non-blocking reads and writes.

Update serial example to use this and remove bindings for uart_pipe.