Remove stage2 sysroot
Apps will need to use path-relative dependencies to get to the zephyr
crates here that are not in the stage 1 sysroot. Some of the stage 2
zephyr crates require commonly-used crates like futures. If those are
included in the sysroot, they will conflict with the version pulled in
by cargo dependencies of the app. Every consumer (which may be an
intermediate dependency from crates.io) would have to be modified to use
"extern crate" and/or remove the dependency from Cargo.toml, which is
not feasible.