Remove intermediate link
The main reason for this was to rename the data/bss sections to put them
in an app mem region for userspace. The zephyr_library_app_memory
feature from Zephyr 1f9723af19af70506118f00db531c2b1dd1db1dd provides a
standard way to do this on a per-library basis.

Unfortunately this feature is not present in 1.14, so it must be
backported if using CONFIG_USERSPACE on that branch.

Remove our custom re-linked library and move the C code to a normal
Zephyr library.