mempool: fail allocations if alignment is not satisfied
Generated code, at least on ARM, does make use of the requested
alignment. Not upholding this assumption can cause even safe Rust code
to clobber the allocator's bookkeeping that comes just before the
beginning of the allocation.

Rather than allocate extra space, assert that the mem pool is configured
so all allocations will be aligned to the max that Rust will ever
request on the platform. On 32-bit ARM, this is empirically 8.