Update syscall bindings for v2.1
Zephyr changed the way syscall headers are generated. Start from the 2.1
version of gen_syscall_header.py and strip it down to create wrappers
and their definitions. All the script has to do is iterate through
syscalls.json and generate a pass-through wrapper function for each.

Build each of syscall-thunk-{any,kernel,user}.c to build with defaults,
__ZEPHYR_SUPERVISOR__, or __ZEPHYR_USER__, respectively. The end result
is concrete implementations for each syscall for each mode.

Simplify zephyr-bindgen to call the wrappers even for kernel mode. This
may result in an unnecessary wrapper for z_impl* functions that are not
inline, but it seems most are inline. The previous method of determining
which functions were inline was seeing which z_impl* functions do not
show up in the bindgen output, because bindgen doesn't generate bindings
for inlines. The new gen_syscall_header in Zephyr creates extern
definitions that overlap with the inline definitions, so this no longer
works. The efficiency is probably not a practical concern, the inlined
syscalls are already optimial, and it makes the generation simpler.

Fixes #16