Add the ability to set cflags on thunks
A specific use-case is when using thumb on arm and the thunks are linked
too far away from the kernel code to use a relative branch. This causes
the linker to generate veneers. But since we are creating wrapper
functions anyway, it would be more efficient for the wrapper to also
serve as the long jump veneer. In this case, the application can compile
with -mlong-calls.

This is highly application and board specific so no policy is here.