uart_buffered: add missing barrier in fifo push/pop
The other end of the lockless fifo runs in IRQ context and is free to
use the offset at the read/write position as soon as it is updated. So
we must guarantee the access to that offset is finished before updating
it.

This is standard stuff from kfifo, but somehow I missed it.