zephyr-futures: allow waking tasks on other threads
An executor and reactor are bound to a single thread, but it should be
possible to have executors on multiple threads and wake tasks that
belong to a different thread.

This would be useful for SMP, or more likely if there are tasks that
need to be given different OS scheduler priorities.

Waking is accomplished by requiring a k_poll_signal for each executor.
The signal is raised by the task waker only if the thread id of the task
is not the current thread.