zephyr-futures: don't register the same event multiple times
Even if a poll implementation tries to only register for the event after
the first time progress can no longer be made, it still has to check
again after registering. If progress can then be made, this series of
events can repeat, theoretically allowing the list of events to grow
unbounded. Make sure the event is not already in the list before
insertion.

Also use reserve_exact() when expanding events. reserve() uses the
"amortized" realloc strategy which grows exponentially. For the embedded
use case, we really want to reserve space for only exactly one more
event.