Add build environment containers and test scripts
We want to support multiple versions. Switching between zephyr versions
is quite time-consuming due to the size and complexity of all the
dependencies (zephyr, sdk, west modules). Build containers for the
configurations we support.

build-all.sh is similar in purpose to Zephyr's sanitycheck, but we have
a fixed set of things to test, so keep things simple for now.