uart-buffered: fix integer promotion in fifo_used()
C's integer promotion would cause the used size to return a bogus large
value if sampled when the u16 write pointer wrapped and the read pointer
had not wrapped. The pointers are intended to wrap naturally. It just
happened that because the pointers were smaller than int, they were
promoted before the subtraction and didn't get truncated as expected
because the size_t return type can hold the large negative int.

This would cause the RX IRQ handler to disable the IRQ when the fifo was
not actually full.