Create a script for generating target json
The json is not stable across rust versions, and all of these are out of
date. It's also very time consuming to handle these manually, given the
number of targets.

This could certainly be generated during the build, but riscv needs to
be handled generically and it would require `jq` in the build
environment.