Always build picolibc from source
Zephyr 3.5 switches to picolibc by default and can use a binary from the
SDK or build from source. Clang, invoked by zephyr-bindgen from
zephyr-sys's build.rs, cannot find the SDK's picolibc include path,
resulting in this error:

> zephyr/include/zephyr/posix/time.h:55:12: fatal error: 'sys/timespec.h' file not found

This is because when using the SDK version, it picks up the include path
from a spec file.

    if(NOT CONFIG_PICOLIBC_USE_MODULE)
    zephyr_compile_options(--specs=picolibc.specs)

We don't pass zephyr_compile_options for bindgen because it contains
many GCC-specific options that fail under clang, and in this case the
spec file won't work with clang anyway.

Build from source so the include path is added in the normal way.