ci: fix build failure in Docker on Zephyr 2.3
Cherry-pick this kconfig bugfix that prevents renaming /dev/null if the
user has permissions to do so. When /dev/null is a real file, rustc
fails with syntax errors.

This is possible because we're running as root in Docker. That could be
revisited but this is the simplest change.

error: failed to run `rustc` to learn about target-specific information

Caused by:
  process didn't exit successfully: `rustc - --crate-name ___
  --print=file-names --crate-type bin --crate-type rlib --crate-type
  dylib --crate-type cdylib --crate-type staticlib --crate-type
  proc-macro --print=sysroot --print=split-debuginfo --print=crate-name
  --print=cfg` (exit status: 1)